fix alignment macro truncation of 64-bit addresses

+++ lib/include/rpmsg_lite.h
@@ -56,9 +56,9 @@ extern "C" {
 /* Shared memory "allocator" parameters */
 #define RL_WORD_SIZE (sizeof(uint32_t))
 #define RL_WORD_ALIGN_UP(a) \
-    (((((uint32_t)a) & (RL_WORD_SIZE - 1U)) != 0U) ? ((((uint32_t)a) & (~(RL_WORD_SIZE - 1U))) + 4U) : ((uint32_t)a))
+    (((((uintptr_t)a) & (RL_WORD_SIZE - 1U)) != 0U) ? ((((uintptr_t)a) & (~(RL_WORD_SIZE - 1U))) + 4U) : ((uintptr_t)a))
 #define RL_WORD_ALIGN_DOWN(a) \
-    (((((uint32_t)a) & (RL_WORD_SIZE - 1U)) != 0U) ? (((uint32_t)a) & (~(RL_WORD_SIZE - 1U))) : ((uint32_t)a))
+    (((((uintptr_t)a) & (RL_WORD_SIZE - 1U)) != 0U) ? (((uintptr_t)a) & (~(RL_WORD_SIZE - 1U))) : ((uintptr_t)a))
 
 /* Definitions for device types , null pointer, etc.*/
 #define RL_SUCCESS    (0)
